cleaned up offset loading code, added sieve offset ratio tracking

+++ src/CUDA/include/sieve_resources.h
@@ -14,7 +14,7 @@ extern uint4 *d_primesInverseInvk[GPU_MAX];
 extern uint32_t *d_primes[GPU_MAX];
 extern uint32_t *d_base_remainders[GPU_MAX];
 extern uint16_t *d_blockoffset_mod_p[GPU_MAX];
-extern uint8_t nOffsetsA;
-extern uint8_t nOffsetsB;
+extern uint32_t nOffsetsA;
+extern uint32_t nOffsetsB;
 
 #endif
